Refactor pipeline module
